Fix missing error when the first character is invalid
In the case where a file started with an invalid character, e.g. a
backtick (`), the parser would incorrectly treat it as an EOF and not
report any errors.

Fixed by checking the lexer for errors immediately after reading from
it in parser::parse.

Added a test called "first_character_invalid". Some changes to the
test.py code were needed to support testing error output.